assignments/parallel-firewall: Add resilience to compilation failures
Make checker look out for the binary to run because if the compiation
failed, there won't be any.

Also, use `$(MAKE)` instead of plain `make` from inside the
`tests/Makefile` as that is the intended way to do it.

Also 2, removing `-i` from the makefile invocation make the inner
makefile fail after the first error, and adding `-` makes it not stop
the outter makefile rule (i.e. also run the checker)

Signed-off-by: Anton-Fabian Patras <patras.fabi@gmail.com>